% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{bre}
\alias{bre}
\title{Specify a branch length (random) effect}
\usage{
bre(
  phyf,
  rate_distribution = c("iid", "laplacian", "student-t", "horseshoe", "Brownian"),
  hyper = list(prec = list(prior = "pc.prec", param = c(1, 0.01))),
  latent = 0
)
}
\arguments{
\item{phyf}{A \code{pfc} column containing the phylogenetic structure}

\item{rate_distribution}{What distribution to use to model rates of evolution?}

\item{hyper}{Hyper parameters as a list. Specify the prior distribution for
\code{engine = INLA} models here.}

\item{latent}{How many latent variables to generate in \code{engine = INLA} models.
Default is none.}
}
\value{
A list of data to be used by the model.
}
\description{
This function is meant to be called only in the \code{formula} argument of \code{fibre()}.
}
