% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{bre_second_order}
\alias{bre_second_order}
\title{Specify a branch length (random) effect for a 'Second Order' Brownian
motion model}
\usage{
bre_second_order(
  phyf,
  hyper = list(prec = list(prior = "pc.prec", param = c(1, 0.01))),
  latent = 0
)
}
\arguments{
\item{phyf}{A \code{pfc} column containing the phylogenetic structure}

\item{hyper}{Hyper parameters as a list. Specify the prior distribution for
\code{engine = INLA} models here.}

\item{latent}{How many latent variables to generate in \code{engine = INLA} models.
Default is none.}
}
\value{
A list of data to be used by the model.
}
\description{
This function is meant to be called only in the \code{formula} argument of \code{fibre()}.
}
