% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ead.R
\name{evo_autodecoder}
\alias{evo_autodecoder}
\title{Create a Evolutionary Autodecoder Model}
\usage{
evo_autodecoder(
  latent_dim,
  n_edges,
  decoder,
  reconstruction_loss,
  device,
  decoder_args = list(),
  loss_args = list()
)
}
\arguments{
\item{latent_dim}{Number of latent dimensions}

\item{decoder}{A \code{torch::nn_model()} specifying a 'decoder' network architecture.
The decoder network should accept a 2 dimensional \code{torch::torch_tensor()} with
first d}
}
\value{
A \code{torch::nn_module()}
}
\description{
Create a Evolutionary Autodecoder Model
}
