% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{get_tips}
\alias{get_tips}
\title{Title}
\usage{
get_tips(
  x,
  type = c("marginals", "samples", "mode", "mean", "median", "lower", "upper", "ci",
    "hpd", "sd"),
  n = 1,
  p = 0.05
)
}
\arguments{
\item{x}{A fitted model object produced by \code{\link{fibrer}}}

\item{type}{What kind of posterior summary to return?}

\item{n}{If \code{type = "samples"}, how many samples to return?}

\item{p}{If \code{type = "hpd"}, what alpha levels to use?}
}
\value{
For all types except "hpd", "ci", and "marginals", a
numeric vector, otherwise a list for "hpd" and "marginals", and a matrix
for "ci".
}
\description{
Title
}
