% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{load_model}
\alias{load_model}
\title{Load a model}
\usage{
load_model(name)
}
\arguments{
\item{name}{Name of the model. Currently only \code{"bird_beak"}.}
}
\value{
A \code{torch::nn_module()} with pre-trained weights
}
\description{
Load a model
}
\examples{
if(torch::torch_is_installed()) {
model <- load_model("bird_beaks")
}
}
