% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_utils.R
\name{make_L}
\alias{make_L}
\title{Make to the root-to-tip matrix of a phylogeny}
\usage{
make_L(
  phy,
  return_nodes = c("tips", "internal", "both"),
  return_type = c("matrix", "list")
)
}
\arguments{
\item{return_nodes}{}
}
\value{

}
\description{
Make to the root-to-tip matrix of a phylogeny
}
