% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_utils.R
\name{make_root2tip}
\alias{make_root2tip}
\title{Make to the root-to-tip matrix of a phylogeny}
\usage{
make_root2tip(
  phy,
  return_nodes = c("tips", "internal", "both"),
  return_type = c("matrix", "list"),
  return_ages = FALSE,
  return_parents = TRUE,
  order = "first",
  sparse = TRUE,
  threads = 1
)
}
\arguments{
\item{phy}{An object of class \code{phylo} containing the phylogeny to generate a root-to-tip
matrix for. Must be a rooted phylogeny but can contain polytomies (that is \code{is.binary(phy)} can
be \code{FALSE}).}

\item{return_nodes}{Which nodes to return rows for? (Can be \code{"tips"}, \code{"internal"} or \code{"both"}).}

\item{return_type}{If \code{return_nodes == "both"}, pass \code{"matrix"} to return a \code{matrix} with
both node types bound together using \code{rbind}, or pass \code{"list"} to return two separate matrices
in a list (with \code{"tips"} as the first element and \code{"internal"} as the second). Ignored if
\code{return_nodes != "both"}.}

\item{return_ages}{Should a vector of node 'ages' be added as an attribute named \code{"ages"} to
the return value? 'ages' is defined as the total distance to the node from the phylogenies
root. Note this is only really an age if the phylogeny is a time tree and that it is the
opposite of the usual meaning of age, that is, it doesn't measure the age of the clade
descending from the node, but how much time (or change accumulated) since the common ancestor
of the whole phylogeny (the root) when this node is reached. This definition of 'age'
makes sense whether the phylogeny is a time-tree or not. If a time-tree the more usual
definition of age can be calculated by subtracted the focal node's age from the age of
the tips (which should all be the same for a time-tree).}

\item{order}{Can be \code{"first_order"} or \code{"second_order"}, specifying whether to return
the root-to-tip matrix in a first order or a second order specification. See documentation
of \code{\link[=fibre]{fibre()}} for a description of the difference.}

\item{sparse}{Should the returned root-to-tip matrix use a sparse matrix representation?
If \code{FALSE} a regular dense matrix will be returned. The sparse matrix representation is
recommended, especially for large trees. Beyond a certain size of tree, only a
sparse return will work, a dense matrix will exhaust all memory. Most root-to-tip
matrices are quite sparse, and so can represent very large trees where dense
matrices cannot (how sparse exactly depends on the shape of the tree).}

\item{return_parent}{If \code{TRUE}, the parent node, matching the root to tip matrix columns,
will be returned as an attribute \code{"parents"}.}
}
\value{
A \code{matrix} if \code{sparse = FALSE}, or a \code{Matrix::dgCMatrix} otherwise.
}
\description{
Make to the root-to-tip matrix of a phylogeny
}
\examples{
test_tree <- ape::rtree(1000)
rtp_mat <- make_root2tip(test_tree)
rtp_mat
}
