% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibre.R
\name{p}
\alias{p}
\title{Phylogenetic Branch Random Effect. To be used in the \code{formula} argument to \code{fibre}.}
\usage{
p(
  id,
  phy = NULL,
  rate_model = c("iid", "gaussian", "ridge", "laplacian", "lasso", "double-exponential",
    "student-t", "horseshoe", "diverging", "fixed", "Brownian"),
  rate_order = c("first", "second"),
  weights = NULL,
  rate_weights = NULL,
  aces = TRUE,
  constr = FALSE,
  hyper = NULL,
  fixed_anc = NULL,
  rate_group = 1L,
  multivariate = 0,
  name = NULL,
  ancestral = FALSE,
  kappa_correction = TRUE
)
}
\arguments{
\item{id}{Character or integer vector with ids to match data to phylogeny. If character,
elements are matched to tip labels in \code{phy}. If integer, they will treated as indexes into
\code{phy$ip.label}. Can also be a \link[ape:read.tree]{ape::phylo} object, in which case the \code{phy} argument is
ignored, and the \code{id} will be an integer indexing the tip labels (in other words, this
works if you know your data is already in the same order as your phylogeny's tip labels).}

\item{phy}{A \link[ape:read.tree]{ape::phylo} object with the phylogeny to be used in the \code{phybre} model.}

\item{fixed_anc}{}
}
\description{
Phylogenetic Branch Random Effect. To be used in the \code{formula} argument to \code{fibre}.
}
