% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phinla.R
\name{fibre}
\alias{fibre}
\title{Phylogenetic Branch Regression (fibre)}
\usage{
fibre(
  formula = ~1,
  phy,
  data = NULL,
  phy_match = "auto",
  family = "gaussian",
  rate_model = "ridge",
  rate_order = c("first_order", "second_order"),
  fit = TRUE,
  aces = TRUE,
  hyper = NULL,
  obs_error = c("est", "one", "zero"),
  ...
)
}
\arguments{
\item{formula}{formula specifying the response (LHS) and set of predictors (RHS) for the
phylogenetic model. The default \code{ ~ 1} is special shorthand specifying a model of
a response with only a phylogenetic model (e.g. no covariates), in which case, the \code{data}
argument must contain a vector or column matrix/data.frame with the response variable(s)
(e.g. species traits).}

\item{phy}{A \code{phylo} object containing the phylogeny to be used for the phylogenetic trait
model.}

\item{data}{A matrix or data.frame (or vector) containing the variables referred to in
\code{formula} argument.}

\item{phy_match}{This argument specifies how the data should be matched to the phylogeny. The
default \code{"auto"}, will match by names or rownames if they exist in \code{data}, or match
by order if no names are present. You can also explicitly specify how to match (recommended).
Option include "names", to match by names (throwing an error if they aren't present), "order", to
match by ordering (element or rows in \code{data} are in the same order as \code{phy$tip.label}),
or any length 1 character vector that refers to a column name in \code{data}, containing a character
vector. In this case, this character vector will be used to match the data to the tip labels in
\code{phy}. This is the recommended best way, as it is the most compatible with "tidy" principles,
making sure species names are treated as data and stay associated with related data.}

\item{family}{Family for the distribution of errors. Default is "gaussian", but can be any
supported by \code{inla}. See \code{names(inla.models()$likelihood)} for a list of possibilities
and use \code{\link[INLA]{inla.doc}} for details of all available families.}

\item{rate_model}{The model to use for the evolutionary rate variation along the phylogeny
(in other words, the choice of prior for rates). Current choices are:
\itemize{
\item{\code{"ridge"}}{Rates are completely independent along all branches but are shrunk
towards zero by an independent Normal prior. This is classic Bayesian Ridge Regression and has
a single hyperparameter whose prior determines the degree of shrinkage.}
\item{\code{"temporal"}}{Rates are constrained to be similar if they are close together
in 'time' (where time is defined as the distance from the root on the phylogeny).
Has a single hyperparameter whose prior determines the degree of temporal
"smoothing". Heavy smoothing forces rates to change slowly through time, light smoothing
allows rates to change quickly (e.g. they can be "wiggly"). Specify a particular model of
temporal autocorrelation with the \code{temporal_model} argument.}
}}

\item{...}{}
}
\value{

}
\description{
Phylogenetic Branch Regression (fibre)
}
