% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nets.R
\name{sdf_net}
\alias{sdf_net}
\title{A signed distance field based neural network model for generating 3d shapes}
\usage{
sdf_net(n_latent = 64, breadth = 512)
}
\arguments{
\item{n_latent}{Number of dimensions for the latent space}

\item{breadth}{Breadth of the multilayer perceptron networks}
}
\value{
A \code{torch::nn_module()}
}
\description{
A signed distance field based neural network model for generating 3d shapes
}
\examples{
if(torch::torch_is_installed()) {
sdf_net()
}
}
