% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{simulate_traits}
\alias{simulate_traits}
\title{Function to simulate continuous trait value histories on a phylogeny.}
\usage{
simulate_traits(
  phy,
  rate_model = c("continuous", "discrete"),
  temp_trend_rates = 0,
  rate_change,
  rates = NULL,
  anc = c(`1` = 0),
  internal = FALSE,
  nsim = 1,
  pos_strat = c("none", "log", "add_const"),
  temp_trend_mean = 0
)
}
\arguments{
\item{phy}{A phylogenetic tree (phylo object) on which to simulate traits}

\item{rate_model}{The type of rate model for how rates of evolution evolve
on the phylogeny: "continuous" for continuous Brownian motion evolution of rates,
or "discrete" for evolution of rate "classes" across the phylogeny, using an mk model.}

\item{temp_trend_rates}{What temporal trend in rates should there be? A positive
number for an increase, and negative number for a decrease with the magnitude controlling
the strength of linear change. This trend is added to rates simulated under the rate_model.}

\item{rate_change}{If \code{rate_model} is "continuous", this should be a single
positive number controlling how fast rates change continuously along the tree. If
\code{rate_model} is "discrete", this should be a transition matrix for the rate classes.
Or, if \code{rate_model} is "discrete", and this can be a length 2 numeric vector specifying}

\item{rates}{Only used if \code{rate_model} is "discrete", in which case this should
be a named vector whose values are the rates in each rate class, and whose names are the
rate class states (e.g. \code{c("1" = 3, "2" = 10)}). See \code{\link[phylotools]{sim.history}},
for more detail on how the discrete model works. Or, if an unnamed numeric vector of length
two, a mean and standard deviation parameterizing a normal distribution from which to
draw rates for each rate class. If \code{NULL}, rates will be drawn from a normals distribution
with mean = 0 and sd = 1.}

\item{anc}{Value of the trait at the root ancestor. For \code{rate_model = "discrete"},
can be a length 1 named vector where the name is the ancestral state, and the value is
the trait starting value. For \code{rate_model = "continuous"}, any names are ignored,
but should be length 2, where the first element is the ancestral trait value and the
second element is the ancestral rate of evolution.}

\item{internal}{Logical value. If \code{TRUE} return trait states at internal nodes.}

\item{nsim}{Number of simulation to run.}

\item{pos_strat}{?}

\item{temp_trend_mean}{A temporal trend in rates.}
}
\value{
A vector or matrix (for \code{nsim > 1}) containing simulated trait values
for each tip if \code{internal = FALSE}, or for each node if \code{internal = TRUE}
}
\description{
Function to simulate continuous trait value histories on a phylogeny.
}
