% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.R
\name{tidy.fibre}
\alias{tidy.fibre}
\title{Tidy Model Results}
\usage{
\method{tidy}{fibre}(
  x,
  effects = c("fixed", "rates", "random", "hyper"),
  conf.type = c("cred.int", "marginals"),
  indexes = NULL,
  ...
)
}
\arguments{
\item{x}{A \code{fibre} model object}

\item{effects}{Which effects do you want tidied? One of: \code{"fixed"},
for fixed effects, \code{"random"} for random effects, or \code{"hyper"} for
the hyper-parameters of the random effects. Can also be \code{"rates"},
which is a synonym for \code{"random"}, since the random effects are
rates of trait evolution along phylogenetic edges.}

\item{conf.type}{What kind of confidence interval. Choices are:
\code{"cred.int"} for approximate Bayesian marginal credible intervals. or
\code{"marginals"} for the full approximate marginal distributions, as a
\code{data.frame} with \code{value} and \code{y.value} columns. \code{value} is the value
of the parameter, and \code{y.value} is the marginal posterior density
(e.g. what \code{value} is the x axis and \code{y.value} is the y axis when plotting
the posterior density).}

\item{indexes}{If \code{effects = "random"} or \code{effects = "rates"}, this is a
vector of indices to retrieve particular random effects. Default is to return
all random effects, however, this can be slow for retrieving the marginals.}

\item{...}{Not used.}
}
\value{
A tidy \code{tibble} with information about the fitted model parameters.
}
\description{
Tidy Model Results
}
