% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibre-fit.R
\name{fibre}
\alias{fibre}
\alias{fibre.default}
\alias{fibre.data.frame}
\alias{fibre.matrix}
\alias{fibre.formula}
\alias{fibre.recipe}
\title{Fit a \code{fibre}}
\usage{
fibre(x, ...)

\method{fibre}{default}(x, ...)

\method{fibre}{data.frame}(x, y, intercept = TRUE, engine = c("inla", "glmnet"), ...)

\method{fibre}{matrix}(x, y, intercept = TRUE, engine = c("inla", "glmnet"), ...)

\method{fibre}{formula}(formula, data, intercept = TRUE, engine = c("inla", "glmnet"), ...)

\method{fibre}{recipe}(x, data, intercept = TRUE, engine = c("inla", "glmnet"), ...)
}
\arguments{
\item{x}{Depending on the context:
\itemize{
\item A \strong{data frame} of predictors.
\item A \strong{matrix} of predictors.
\item A \strong{recipe} specifying a set of preprocessing steps
created from \code{\link[recipes:recipe]{recipes::recipe()}}.
}}

\item{...}{Not currently used, but required for extensibility.}

\item{y}{When \code{x} is a \strong{data frame} or \strong{matrix}, \code{y} is the outcome
specified as:
\itemize{
\item A \strong{data frame} with 1 numeric column.
\item A \strong{matrix} with 1 numeric column.
\item A numeric \strong{vector}.
}}

\item{formula}{A formula specifying the outcome terms on the left-hand side,
and the predictor terms on the right-hand side.}

\item{data}{When a \strong{recipe} or \strong{formula} is used, \code{data} is specified as:
\itemize{
\item A \strong{data frame} containing both the predictors and the outcome.
}}
}
\value{
A \code{fibre} object.
}
\description{
\code{fibre()} fits a model.
}
\examples{
predictors <- mtcars[, -1]
outcome <- mtcars[, 1]

# XY interface
mod <- fibre(predictors, outcome)

# Formula interface
mod2 <- fibre(mpg ~ ., mtcars)

# Recipes interface
library(recipes)
rec <- recipe(mpg ~ ., mtcars)
rec <- step_log(rec, disp)
mod3 <- fibre(rec, mtcars)

}
