% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibre-fit.R
\name{fibre}
\alias{fibre}
\alias{fibre.default}
\alias{fibre.data.frame}
\alias{fibre.matrix}
\alias{fibre.formula}
\alias{fibre.recipe}
\title{Fit a \code{fibre}}
\usage{
fibre(x, ...)

\method{fibre}{default}(x, ...)

\method{fibre}{data.frame}(
  x,
  y,
  intercept = TRUE,
  engine = c("inla", "glmnet", "torch"),
  engine_options = list(),
  ncores = NULL,
  verbose = 0,
  fit = TRUE,
  ...
)

\method{fibre}{matrix}(
  x,
  y,
  intercept = TRUE,
  engine = c("inla", "glmnet", "torch"),
  engine_options = list(),
  ncores = NULL,
  verbose = 0,
  fit = TRUE,
  ...
)

\method{fibre}{formula}(
  formula,
  data,
  intercept = TRUE,
  family = "gaussian",
  engine = c("inla", "glmnet", "torch"),
  engine_options = list(),
  ncores = NULL,
  verbose = 0,
  fit = TRUE,
  ...
)

\method{fibre}{recipe}(
  x,
  data,
  intercept = TRUE,
  engine = c("inla", "glmnet", "torch"),
  engine_options = list(),
  ncores = NULL,
  verbose = 0,
  fit = TRUE,
  ...
)
}
\arguments{
\item{x}{Depending on the context:
\itemize{
\item A \strong{data frame} of predictors.
\item A \strong{matrix} of predictors.
\item A \strong{recipe} specifying a set of preprocessing steps
created from \code{\link[recipes:recipe]{recipes::recipe()}}.
}}

\item{...}{Not currently used, but required for extensibility.}

\item{y}{When \code{x} is a \strong{data frame} or \strong{matrix}, \code{y} is the outcome
specified as:
\itemize{
\item A \strong{data frame} with 1 numeric column.
\item A \strong{matrix} with 1 numeric column.
\item A numeric \strong{vector}.
}}

\item{intercept}{A logical. Should an intercept be included in the model?}

\item{engine}{A single character. The engine to use for fitting the model.}

\item{engine_options}{A list of options to pass to the engine.}

\item{ncores}{An integer. The number of cores to use for parallel processing.}

\item{verbose}{An integer. The level of verbosity.}

\item{fit}{A logical. Should the model be fit? If \code{FALSE}, the model is not fit
and instead a list of data is returned that can be used to fit the model later.
Useful for debugging or for fitting the model using a custom model design or
currently unsupported engine.}

\item{formula}{A formula specifying the outcome terms on the left-hand side,
and the predictor terms on the right-hand side.}

\item{data}{When a \strong{recipe} or \strong{formula} is used, \code{data} is specified as:
\itemize{
\item A \strong{data frame} containing both the predictors and the outcome.
}}
}
\value{
A \code{fibre} object.
}
\description{
\code{fibre()} fits a model.
}
\examples{
predictors <- mtcars[, -1]
outcome <- mtcars[, 1]

# XY interface
#mod <- fibre(predictors, outcome)

# Formula interface
#mod2 <- fibre(mpg ~ ., mtcars)

# Recipes interface
#library(recipes)
#rec <- recipe(mpg ~ ., mtcars)
#rec <- step_log(rec, disp)
#mod3 <- fibre(rec, mtcars)

}
