% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibre.R
\name{fibre_data}
\alias{fibre_data}
\title{Generate data formatted to fit a \code{fibre} model.}
\usage{
fibre_data(formula = ~1, data = NULL)
}
\arguments{
\item{formula}{Formula specifying a \code{fibre} model see details for more information}

\item{data}{A data.frame or named list containing variables referenced in \code{formula}.
The only variable that doesn't have to be in \code{data} is a \code{phylo} representing a
phylogeny to be used in the \code{\link[=p]{p()}} function.}
}
\value{
A \code{fibre_data} object containing all data necessary to fit a fibre model using
\code{\link[=fibre_fit]{fibre_fit()}}.
}
\description{
Generate data formatted to fit a \code{fibre} model.
}
