% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{re}
\alias{re}
\title{Specify a random effect}
\usage{
re(groups, hyper = list(prec = list(prior = "pc.prec", param = c(1, 0.01))))
}
\arguments{
\item{groups}{A character or factor column containing the grouping variable
for the random effect}

\item{hyper}{Hyper parameters as a list. Specify the prior distribution for
\code{engine = INLA} models here.}

\item{latent}{How many latent variables to generate in \code{engine = INLA} models.
Default is none.}
}
\value{
A list of data to be used by the model.
}
\description{
This function is meant to be called only in the \code{formula} argument of \code{fibre()}.
}
